Add hidden CreateUser endpoint to TokenController
A new `CreateUser` endpoint was added to the `TokenController`
in the `CleanArchMvc.API.Controllers` namespace. This endpoint
handles HTTP POST requests to register a new user using the
`_authentication.RegisterUser` method.

If registration succeeds, it returns an HTTP 200 OK response
with a success message. If registration fails, it returns an
HTTP 400 Bad Request response with validation errors.

The endpoint is decorated with `[ApiExplorerSettings(IgnoreApi = true)]`
to hide it from Swagger documentation.